refactor: extract chatgpt room with two components(#28)
